GHC 8.0 compat
Backwards incompatible changes in GHC API unfortunately require
CPP to handle.

GHC.SrcLoc was removed and some of it moved to GHC.Stack

GHC.Stack is now imported with explicit name list to solve conflict with
prettyCallStack function. And it is a good thing to do with unstanble
GHC API in general.

WithCallstack lost Eq instance because CallStack also lost it in ghc-8.0
and this change requires major version bump.